requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
opencv-python-headless==4.10.0.84
openai==1.55.3
python-dotenv==1.0.1
requests==2.32.3
